test: 테스트 코드 리팩터링
